splited marc bib edit into two files. removed dependenceies between tests
